Account Creation and Login UI + authentication checking with sessions.
I have now integrated an API, which can be called upon, to check whether the user has logged in. In other words, if a session has been created. Additionally, I have added a logout button within the navigation bar on the marketplace. Once pressed, the user will be redirected to the customerAccountManagement page to log back in. Both of these features ensure users must log in to create a valid session which in return allows them to view the marketplace.