Significant Backend adjustments
I have adjusted the backend significantly, meaning all users are initially "customers" and can then decide to be a "business". This means the user can access the business dashboard.

I have also introduced 2 new schemas. This includes; Listings and Bookings. Listings is used to hold the service cards the businesses create and Bookings is used to hold the individuals customer bookings.