Bookings View + Navigation Refinement
I have implemented a bookings view where all the bookings are fetched together. This still needs enhancing to improve user feedback, but allows the user to view all the bookings associated to their account.

I have also improved the navigation bar to enhance the ease of navigation throughout the application. This uses switch cases for the button text depending on the users URL and a basic if-else sequence for the buttons redirection depending on the current URL

Finally, I have introduced loading animations to the marketplace to improve the user feedback. Prior to this change, the display would be empty until the destinations were fetched.