Very Basic Price Calculator:
I have implemented a very basic price calculator to update the TotalPrice within the UI upon change in a variable, such as the num of adults. (This is accessed in the UseState)

In later iterations, the price will be determined by num of adults and children (with varied pricing) and length of stay. I could also add "extras"