Image Saving Rework
I had come into issues with the data size of the images, resulting in large marketplace loading times as the whole image would need to be fetched from the MongoDb database. To address this, I am now using Cloudiary to store the images and I access the image via the URL, which is stored in the mongoDB database listing document. This process is carried out through the use of Cloudaries SDK