Early Analytics Page design.
I have developed the early version of the business analytics. Whilst this has taken a long time to figure out how to use the mongoDb aggregate fetching with my relational MongoDb database, I have now created an API that fetches the total revenue generated from the users listings. additionally, the Api formats this into weeks using the booking document createdAt attribute.

In following iterations, I would like to allow the user to input specific listings to view its analytics. This may be quite complex as the API will need adjusting / additional Apis created. Maybe using the input listing as the match operator filter, rather than the users document id. This is used because it's the same as the businessId on the listing.

PS - session attributes may need to be defined as an objectId before using within the MongoDb Aggregate fetch operators. Errors will occur if not done so.